requirements.txt
discord.py
aiomysql
ormar
uvloop